requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
bs4==0.0.2
build==1.4.3
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.7
click==8.3.2
cmudict==1.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.7
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
Distance==0.1.3
distro==1.9.0
durationpy==0.10
emoji==1.2.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.135.3
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
fpdf==1.7.2
frozenlist==1.8.0
fsspec==2026.3.0
gast==0.7.0
geopandas==1.1.3
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.2
google-genai==1.73.1
google-pasta==0.2.0
googleapis-common-protos==1.74.0
gradio==6.12.0
gradio_client==2.4.1
greenlet==3.4.0
groovy==0.1.2
grpcio==1.80.0
gTTS==2.5.4
h11==0.16.0
h5py==3.14.0
hangul-jamo==1.0.1
hf-gradio==0.3.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==7.1.0
iniconfig==2.3.0
ipykernel==7.2.0
ipython==8.39.0
jamo==0.4.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
JPype1==1.4.1
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keras==3.12.1
kiwisolver==1.5.0
kollocate==0.0.2
konlpy==0.6.0
koparadigm==0.10.0
kss==6.0.6
kubernetes==35.0.0
langchain==1.2.15
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
libclang==18.1.1
llvmlite==0.47.0
lxml==6.0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.1
more-itertools==11.0.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.19.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.65.0
numpy==2.2.6
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.23.2
openai-whisper==20250625
opencv-python==4.13.0.92
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opt_einsum==3.4.0
optree==0.19.0
orjson==3.11.8
ormsgpack==1.12.2
osmnx==2.0.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
parso==0.8.6
pecab==1.0.8
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.20.0
pyogrio==0.12.1
pyparsing==3.3.2
PyPika==0.51.1
pyproj==3.7.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytorch-tabnet==4.1.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.4.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
tf_keras==2.21.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.1
tomlkit==0.14.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tossi==0.3.1
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.1
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.37
ultralytics-thop==2.0.18
Unidecode==1.4.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Whoosh==2.7.4
wordcloud==1.9.6
wrapt==2.1.2
xlrd==1.2.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
zstandard==0.25.0
